moved sync stuff into its own file, mpd stuff into mpdfactory file
